fix: move plant and integrator for VIMPPI to designated folders
