+++ src/python/double_pendulum/model/jax_plant.py
